% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_color_schema}
\alias{get_color_schema}
\title{Get colors for a discrete variable}
\usage{
get_color_schema(labels, grayscale = F)
}
\arguments{
\item{labels}{The input variable}

\item{grayscale}{If TRUE returns grayscale colors}
}
\description{
Get different colors for a vector, for example a Treatment variable
}
\author{
BDA Team \email{matias.thayer@owlstone..co.uk}
BDA Team
License: GNU GPL (>= 2)
}
