% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{get_factor_at_time_to_divide_by}
\alias{get_factor_at_time_to_divide_by}
\title{Obtain Owlstone scalling factors}
\usage{
get_factor_at_time_to_divide_by(df, time_col, feature_cols)
}
\arguments{
\item{df}{Input data frame (rows as samples and columns as features)}

\item{time_col}{column that contains the time aggregation (e.g. sequence or batch)}

\item{feature_cols}{vector with the name of the features to use}
}
\description{
Normalise data assuming that samples are in rows, and molecular features in columns
}
\author{
BDA Team \email{matias.thayer@owlstone..co.uk}
BDA Team
License: GNU GPL (>= 2)
}
