% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{norm_owlstone}
\alias{norm_owlstone}
\title{Scale data frame using scaling factor (divide by scalling factors)}
\usage{
norm_owlstone(df_to_scale, time_col, feature_cols, factor_at_time_to_divide_by)
}
\arguments{
\item{df_to_scale}{Input data frame (rows as samples and columns as features)}

\item{time_col}{column that contains the time aggregation (e.g. sequence or batch)}

\item{feature_cols}{vector with the name of the features to use}

\item{factor_at_time_to_divide_by}{data frame with the scaling factor. It must be consistent with the time_col parameter}
}
\description{
Normalise data assuming that samples are in rows, and molecular features in columns
}
\author{
BDA Team \email{matias.thayer@owlstone..co.uk}
BDA Team
License: GNU GPL (>= 2)
}
