% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_by_feat}
\alias{scale_by_feat}
\title{Scale data by centering to the mean}
\usage{
scale_by_feat(df, mf_cols, type = "MeanCenter")
}
\arguments{
\item{df}{Input data frame}

\item{mf_cols}{Input name of molecular features columns}

\item{type}{Input name of molecular features columns}
}
\description{
Scale data by column (assuming that samples are in rows, and molecular features in columns)
}
\author{
BDA Team \email{matias.thayer@owlstone..co.uk}
BDA Team
License: GNU GPL (>= 2)
}
